Editor window improvements:
The editor no longer automatically checks every file in OnEnabled. This is called every time you enter play mode, and if you have a lot of scripts, it can cause all kinds of problems.
When checking files, a progress bar appears if checking starts taking longer than a fraction of a second.
Unchecked files now have an image in the list.
Reworked the code to use scopes for vertical layout and scroll views.