STARTING to make it possible to dump pure JSON syntax tree.
Right now this is just stubs for command line options and a tiny bit of
refactoring.

This is going to require refactoring to make it nice.

Basically my plan is to pull the Dialogue out of the Loader, since
the only thing the Loader needs from it are the Logger objects.  Those can
just be passed to it directly.

THEN we are going to make it possible to just get the plain parsed nodes
out of the Loader directly, without requiring the Dialogue itself get involved,
since really what that is for is to RUN the dialogue.

Once that happens then just dumping the syntax tree directly becomes much easier,
and we can also pull showTokens and onlyConsiderNode options out of the Dialogue
object since those only ever get passed through to the Loader.

Whew!