Remove old code for running 'pending' options
Prior to the replacement of [[options]] with the <<jump>> command,
options were added to a list and shown when the node was exited with any
options that had been added since the last AddOptions instruction.
Because all options are now explicitly shown via the -> syntax, this
code is now irrelevant.