Interpolated lines with just vars now work.
Previous modification to allow for interpolated lines to have any old thing meant that as a side-effect when you just had a variable inside a line it wouldn't be type checked.
Fixing this required adding a new tier of diagnostics which are essentially potential diagnostics.
These potential concerns essentially are flagging the maybe no-type issues of variables and are later checked by the compiler to see if they are resolved.
If they are resolved they are ignored, otherwise they are essentially promoted to a full diagnostic.
This isn't ideal but works fine for now, a more all encompassing change to the type checker would be needed to make the system more flexible and less dependant on custom modifications like this.