line tagger better handles emoji.
The previous system uses the text replacements code that relies on determining character offsets in the string to modify the yarn.
This works fine but ANTLR says multiwidth symbols (like emoji) are a single character long.
This means each emoji or similar multiwidth symbol in the yarn is introducing an off by one error of the replacement positions.
Changed it so that the tagger instead now uses a stream rewriter.
Fixes #323 and fixes https://github.com/YarnSpinnerTool/YarnSpinner-Unity/issues/177