moved WS onto hidden channel.
Because the WS was being skipped when the rewriter had to put comments into the output they were being smooshed hard up against the line tag which is invalid.
By keeping the WS but just hiding it the rewriter can use it to correctly keep the spacing as set by the writer.