Updated the tagging test.
There a chance for shared textual lines but with different line tags (one manually added, one curated) to be skipped being checked depending on their ordering in the string table.
Modified the test Yarn for this test so that this happens, and then also fixed the test by running through the known tagged lines firs before the generated lines.
Fixes #363