Make the type solver incremental Previously, the type solver waited until all files were parsed, and then solved all type constraints at once. This meant that handling type solver failure was complex, and frequently meant that the solver wasn't able to determine which errors to produce (and as a result, produced more errors than was actually useful.) Additionally, the solver was difficult to debug.
This change causes the type checker to attempt to resolve and dispose of
as many type constraints as it can after every statement. It does this
by incrementally building up the type solution as it goes; after
successfully resolving the current set of type constraints, the solver
eliminates any constraints that are now fully resolved by the current
solution. This reduces the number of constraints that must be considered
at any one time, increasing performance.

If a statement results in a failure to resolve the current type
constraints, the diagnostics for the failing constraints are logged, and
the solver continues. This reduces the total number of constraints that
may be failing at any one time, reducing error noise.